Make Predictions

Predicts outcome from features with stacked model.

# S3 method for joinet
predict(object, newx, type = "response", ...)

Arguments

object
joinet object

newx
covariates: numeric matrix with \(n\) rows (samples) and \(p\) columns (variables)

type
character "link" or "response"

...
further arguments (not applicable)

Value

This function returns predictions from base and meta learners. The slots base and meta each contain a matrix with \(n\) rows (samples) and \(q\) columns (variables).

Examples

n <- 50; p <- 100; q <- 3
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
Y <- replicate(n=q,expr=rnorm(n=n,mean=rowSums(X[,1:5])))
Y[,1] <- 1*(Y[,1]>median(Y[,1]))
object <- joinet(Y=Y,X=X,family=c("binomial","gaussian","gaussian"))
predict(object,newx=X)

#> $base
#>            [,1]         [,2]       [,3]
#>  [1,] 0.8880859  4.167470686  3.9891323
#>  [2,] 0.8522805  5.339566777  4.0083957
#>  [3,] 0.1409537 -1.127906721 -1.1297818
#>  [4,] 0.5482139  0.509438091  0.3658942
#>  [5,] 0.7212955  0.304478229  0.9022886
#>  [6,] 0.7586983  2.504868432  2.1100457
#>  [7,] 0.1873125 -3.108111939 -2.2772250
#>  [8,] 0.2306623 -1.070706980 -0.3363343
#>  [9,] 0.1047830 -2.822495027 -3.4160136
#> [10,] 0.7379560  0.465976049  0.6822625
#> [11,] 0.8896766  1.158775734  0.4838729
#> [12,] 0.8359719  0.781709813  1.6460842
#> [13,] 0.1160376  0.680443196  0.6246262
#> [14,] 0.6756877  1.781633314  1.8216938
#> [15,] 0.8400620  5.825804542  4.7541491
#> [16,] 0.5643016 -0.091714995 -0.2942564
#> [17,] 0.1739011 -2.563509736 -2.6076003
#> [18,] 0.9466982  5.526587288  5.9721775
#> [19,] 0.1686037 -1.704229560 -2.1406879
#> [20,] 0.1921496 -1.132857441 -1.2162503
#> [21,] 0.2366729 -1.240646242 -1.9384757
#> [22,] 0.1819760 -0.052046078  0.6508637
#> [23,] 0.7250998  0.735685728  2.2780094
#> [24,] 0.8407405  2.607088261  2.2540088
#> [25,] 0.8679057  2.051845998  1.3588967
#> [26,] 0.4135769  2.218927521  1.5887495
#> [27,] 0.5025495  1.128367948  0.4530161
#> [28,] 0.1376866 -1.185377966 -0.2057633
#> [29,] 0.1305300 -2.223211757 -1.8345859
#> [30,] 0.2759690 -1.538698057 -1.4417756
#> [31,] 0.1134478 -3.586361733 -3.8450145
#> [32,] 0.7491258  2.696348673  2.0159418
#> [33,] 0.1268086 -2.895249987 -3.4780149
#> [34,] 0.7312275 -0.169239334  0.6057204
#> [35,] 0.8466929  4.118935091  3.2971451
#> [36,] 0.6942457 -0.489125471  1.0074404
#> [37,] 0.3977223  1.241781415  1.8756228
#> [38,] 0.2910864 -1.581735803 -1.9666797
#> [39,] 0.9709208  2.429454400  3.0653045
#> [40,] 0.0976876  0.158660992 -0.1079414
#> [41,] 0.2021189 -2.971975453 -2.0492191
#> [42,] 0.4597892  0.079612742  0.6544148
#> [43,] 0.8410031  1.948914331  1.6881881
#> [44,] 0.9088317  2.917545063  2.6839440
#> [45,] 0.6070912  0.005819128  0.4001448
#> [46,] 0.1247350 -4.043725444 -3.2091078
#> [47,] 0.2831481 -1.312531458 -1.3224245
#> [48,] 0.8381451  3.016884385  3.2377296
#> [49,] 0.4860702 -0.216995747 -0.4668990
#> [50,] 0.3440635 -1.076599335 -0.5240837
#> 
#> $meta
#>             [,1]        [,2]       [,3]
#>  [1,] 0.98064463  5.19897697  4.7360747
#>  [2,] 0.99197937  5.84366962  5.5963903
#>  [3,] 0.18943662 -1.51046356 -1.2476864
#>  [4,] 0.54641174  0.49838768  0.5662083
#>  [5,] 0.54354221  0.79293031  0.6407069
#>  [6,] 0.89693547  2.88724445  2.7440331
#>  [7,] 0.03697401 -3.44088916 -3.1657704
#>  [8,] 0.23183479 -0.88101613 -0.8754356
#>  [9,] 0.03394474 -4.14660385 -3.4326588
#> [10,] 0.55921306  0.71373646  0.6664470
#> [11,] 0.67025077  0.93651367  1.0875391
#> [12,] 0.67589797  1.61073592  1.2976915
#> [13,] 0.59505302  0.78555158  0.7983773
#> [14,] 0.82331123  2.28083377  2.0979672
#> [15,] 0.99543790  6.66779448  6.2607396
#> [16,] 0.39040527 -0.32292796 -0.1460278
#> [17,] 0.05040084 -3.39737740 -2.9074431
#> [18,] 0.99586835  7.42598745  6.5507606
#> [19,] 0.10334833 -2.58309179 -2.0880397
#> [20,] 0.18534147 -1.57836991 -1.2873209
#> [21,] 0.14735537 -2.18127513 -1.6666767
#> [22,] 0.45944323  0.41153872  0.2766830
#> [23,] 0.70460008  2.06281035  1.5275712
#> [24,] 0.90716890  3.05082839  2.8783567
#> [25,] 0.83474192  2.07690598  2.1015877
#> [26,] 0.85907458  2.34016966  2.3188687
#> [27,] 0.66330123  0.89688254  1.0525690
#> [28,] 0.22272870 -0.84414480 -0.9044048
#> [29,] 0.07788057 -2.63114478 -2.3378530
#> [30,] 0.13608854 -1.96673313 -1.6764118
#> [31,] 0.01726633 -4.88098590 -4.1668834
#> [32,] 0.90749211  2.91918450  2.8440511
#> [33,] 0.03167131 -4.23250230 -3.5114006
#> [34,] 0.43453178  0.31446891  0.1726533
#> [35,] 0.97594969  4.65074439  4.4124062
#> [36,] 0.40179114  0.44560499  0.1074598
#> [37,] 0.75854279  2.03128364  1.7278889
#> [38,] 0.11698940 -2.38593747 -1.9264531
#> [39,] 0.91390288  3.56748886  3.0872866
#> [40,] 0.44863239 -0.04773375  0.1136769
#> [41,] 0.04328189 -3.19565607 -2.9717607
#> [42,] 0.48451330  0.48500229  0.3738984
#> [43,] 0.83614778  2.27003302  2.1639687
#> [44,] 0.93274581  3.54214729  3.2832743
#> [45,] 0.45365998  0.25346926  0.2142760
#> [46,] 0.01452023 -4.64705697 -4.2344494
#> [47,] 0.16178510 -1.75508226 -1.4622172
#> [48,] 0.94547843  4.01341376  3.5862912
#> [49,] 0.35740116 -0.52055030 -0.3090718
#> [50,] 0.22223342 -1.02585015 -0.9579621
#>