Make Predictions

Predicts outcome from features with stacked model.

# S3 method for joinet
predict(object, newx, type = "response", ...)

Arguments

object
joinet object

newx
covariates: numeric matrix with \(n\) rows (samples) and \(p\) columns (variables)

type
character "link" or "response"

...
further arguments (not applicable)

Value

This function returns predictions from base and meta learners. The slots base and meta each contain a matrix with \(n\) rows (samples) and \(q\) columns (variables).

Examples

n <- 50; p <- 100; q <- 3
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
Y <- replicate(n=q,expr=rnorm(n=n,mean=rowSums(X[,1:5])))
object <- joinet(Y=Y,X=X)
predict(object,newx=X)

#> $base
#>             [,1]       [,2]         [,3]
#>  [1,] -2.6382210 -2.9659453 -2.306391444
#>  [2,]  2.6552680  1.4525574  2.505892331
#>  [3,]  1.1295699  1.4181935  1.024506708
#>  [4,]  0.2566896 -1.1868062  0.770974904
#>  [5,]  0.9235359  2.1951153  1.435281993
#>  [6,]  0.7376982  1.7955112  1.613380715
#>  [7,] -0.3904051 -0.9624476 -1.305940614
#>  [8,] -4.8823206 -5.1329853 -5.632617590
#>  [9,] -2.8826967 -4.2866886 -3.319082823
#> [10,]  2.4610522  1.7177360  2.029169444
#> [11,]  0.9047337  1.7704820  0.138986055
#> [12,] -3.4383168 -3.4492442 -4.077404028
#> [13,]  0.3573390  1.0892937  0.751307437
#> [14,]  2.6856775  4.1461204  4.109918838
#> [15,] -2.5859515 -3.0620775 -3.039261367
#> [16,]  1.7188957  1.1177316  2.666496452
#> [17,] -2.8405784 -4.3156935 -3.265928842
#> [18,]  2.5205082  3.2674372  4.021626346
#> [19,]  2.4668067  2.2547918  1.521037026
#> [20,] -3.2163874 -5.1967661 -3.914674437
#> [21,] -1.8090909 -2.4841106 -2.800940136
#> [22,]  0.1219993 -0.4755235 -1.346081912
#> [23,]  0.5620446  2.9490435  0.834109773
#> [24,]  0.7231336 -0.8008719  0.006685054
#> [25,]  0.3882856 -0.4834631 -0.389710153
#> [26,]  0.8986010  0.3555128  0.582610989
#> [27,]  0.9810074  0.5822080 -0.083507593
#> [28,]  1.3495727  0.2573044  0.950629861
#> [29,]  1.5941948  1.6883281  2.610167833
#> [30,]  0.5201647 -0.4452296 -1.053437483
#> [31,]  2.5336520  3.9861323  3.091867020
#> [32,]  1.2177945  0.4127256  1.337044870
#> [33,]  2.3160903  2.8285159  2.868384803
#> [34,] -0.9393410 -0.3716184  0.146997363
#> [35,]  2.0302346  2.9987535  2.602339175
#> [36,]  2.5560019  4.2112428  2.946193469
#> [37,]  1.5872347  0.8023620  2.234054545
#> [38,]  1.1985388  1.9795947  1.730649774
#> [39,] -0.2143121  1.3476443 -0.594151486
#> [40,]  1.1493740  0.7250426  0.991941924
#> [41,] -1.8056163 -2.1610821 -2.553768773
#> [42,]  1.4862873  2.8879993  2.480084861
#> [43,]  0.4224656 -0.4430081 -0.643021999
#> [44,] -2.7067256 -2.7823989 -3.502945445
#> [45,]  5.9534870  5.9664792  5.997630680
#> [46,] -0.3503043 -0.5702642 -1.399712343
#> [47,] -2.3705833 -3.7251487 -3.201279704
#> [48,]  2.3128394  0.3872560  2.581339973
#> [49,]  1.9162032  1.5412281  2.499116776
#> [50,]  0.3627602 -0.0460838 -0.163484167
#> 
#> $meta
#>              [,1]        [,2]        [,3]
#>  [1,] -2.90192230 -3.02584785 -3.50908027
#>  [2,]  2.39542048  2.26318849  2.57981667
#>  [3,]  1.37995658  1.19161184  1.36859232
#>  [4,] -0.10442969 -0.23788611 -0.32773037
#>  [5,]  1.84127467  1.54923364  1.76924250
#>  [6,]  1.68959674  1.38037917  1.55660905
#>  [7,] -1.02530473 -1.06748357 -1.20876200
#>  [8,] -5.80753317 -5.79847237 -6.65125141
#>  [9,] -3.96775247 -3.96123175 -4.56926765
#> [10,]  2.26723347  2.13877182  2.45557415
#> [11,]  1.11434304  0.95739165  1.13565036
#> [12,] -4.05024482 -4.09067752 -4.68827456
#> [13,]  0.92384601  0.68180221  0.76833239
#> [14,]  4.25942013  3.84395646  4.37299303
#> [15,] -3.23289184 -3.28042333 -3.77139025
#> [16,]  2.07323424  1.82898732  2.04401186
#> [17,] -3.94797334 -3.93937069 -4.54568065
#> [18,]  3.79066368  3.40995693  3.85973520
#> [19,]  2.29800875  2.18617512  2.53967040
#> [20,] -4.70238788 -4.63546058 -5.34255615
#> [21,] -2.67936429 -2.68297837 -3.06494899
#> [22,] -0.69449828 -0.69598034 -0.75890084
#> [23,]  1.83606495  1.51171990  1.75323104
#> [24,] -0.12829775 -0.15732907 -0.18502812
#> [25,] -0.23621888 -0.28879849 -0.32362810
#> [26,]  0.66711512  0.54654527  0.62279211
#> [27,]  0.51467298  0.45091594  0.54623784
#> [28,]  0.89040074  0.79843332  0.90747793
#> [29,]  2.27148744  1.98600271  2.23240937
#> [30,] -0.45869532 -0.43838407 -0.46392424
#> [31,]  3.73043085  3.39471342  3.89234877
#> [32,]  1.08474130  0.93394585  1.04612756
#> [33,]  3.06877069  2.78831480  3.18080603
#> [34,] -0.30539364 -0.57573031 -0.70918733
#> [35,]  2.96169567  2.66017932  3.03963484
#> [36,]  3.77614022  3.44391387  3.95904729
#> [37,]  1.72151260  1.51618343  1.69400858
#> [38,]  1.93753572  1.66506447  1.89406000
#> [39,]  0.33845084  0.13057382  0.17943816
#> [40,]  1.06385363  0.91822780  1.04543490
#> [41,] -2.43330677 -2.47575048 -2.83188930
#> [42,]  2.72299589  2.37093229  2.69608526
#> [43,] -0.31377697 -0.34193191 -0.37265060
#> [44,] -3.33047920 -3.36708489 -3.85060303
#> [45,]  6.68096086  6.40498435  7.35340901
#> [46,] -0.87950676 -0.92959495 -1.03920314
#> [47,] -3.53887755 -3.50967340 -4.03278950
#> [48,]  1.86583928  1.74202850  1.95196853
#> [49,]  2.24286807  2.01498151  2.27657781
#> [50,]  0.04456881 -0.05474787 -0.05785603
#>