Make Predictions

Predicts outcome from features with stacked model.

# S3 method for joinet
predict(object, newx, type = "response", ...)

Arguments

object
joinet object

newx
covariates: numeric matrix with \(n\) rows (samples) and \(p\) columns (variables)

type
character "link" or "response"

...
further arguments (not applicable)

Value

This function returns predictions from base and meta learners. The slots base and meta each contain a matrix with \(n\) rows (samples) and \(q\) columns (variables).

Examples

n <- 50; q <- 3; p <- 100
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
Y <- replicate(n=q,expr=rnorm(n=n,mean=rowSums(X[,1:5])))
object <- joinet(Y=Y,X=X)
predict(object,newx=X)

#> $base
#>             [,1]        [,2]         [,3]
#>  [1,] -0.6800725 -1.21872520 -0.310657961
#>  [2,]  2.5231822  2.01759432  2.319092565
#>  [3,]  3.0147246  2.95961835  3.071353923
#>  [4,] -2.0321731 -1.16697865 -1.131585864
#>  [5,]  1.1989575  1.94542159  1.205214655
#>  [6,] -5.2181676 -4.73651108 -5.106512539
#>  [7,]  2.6630929  2.28298777  1.837353182
#>  [8,]  4.1517740  3.68154548  2.859426063
#>  [9,]  2.2255938  2.62966728  2.209039320
#> [10,] -0.1976240 -0.26331544 -0.084736597
#> [11,] -2.2848625 -2.88205673 -3.371158126
#> [12,]  0.1395562  0.12678281  1.952331938
#> [13,]  0.5319628  1.08030034  0.877379283
#> [14,] -3.2092512 -3.20808021 -5.574818243
#> [15,]  2.1373104  1.54594741  0.792750994
#> [16,]  1.3677467  0.24229789  1.303661371
#> [17,]  2.4203610  1.97149218  2.164521950
#> [18,] -1.6439085 -1.13752492 -0.148270982
#> [19,] -2.6971069 -3.21077918 -3.448781770
#> [20,]  0.4847756  1.32920669  1.045263833
#> [21,]  1.0580634  1.41141884  0.487897356
#> [22,] -0.1159397  0.33252059  1.764586212
#> [23,] -2.4576322 -2.61545228 -1.960495924
#> [24,] -2.8963417 -2.86185997 -3.306570648
#> [25,]  1.0818188  1.14369892  2.186762596
#> [26,] -3.9271814 -3.84456309 -5.731322146
#> [27,] -2.3652352 -2.86340632 -3.227402032
#> [28,] -2.1800217 -1.48891855 -2.572985868
#> [29,]  0.7460273 -0.27982198 -0.506445835
#> [30,] -0.8429611  0.13194537 -0.737426505
#> [31,] -0.7316201 -1.51707620 -1.153890045
#> [32,] -2.5021179 -1.84304055 -3.478808874
#> [33,]  0.8250113 -0.21440315  0.524465823
#> [34,]  3.0831834  2.23039182  3.727105270
#> [35,]  0.4993630 -1.34134635 -1.164539640
#> [36,]  1.6239756  0.78835676  0.528243360
#> [37,]  2.1683663  0.91851657  1.521169026
#> [38,] -0.9397605  0.08426956 -0.751842377
#> [39,] -0.2803268 -1.20184123 -0.560565492
#> [40,]  1.9729289  3.26803044  2.090648484
#> [41,]  0.7398532 -0.83577871  1.020203057
#> [42,] -1.7847439 -2.02797038 -3.409427348
#> [43,]  0.9128885  1.60609986  2.616774138
#> [44,]  1.1536491  2.16601146 -0.015506128
#> [45,]  0.6315267  0.22276528  0.985103241
#> [46,]  5.4299519  5.38132076  6.438669837
#> [47,]  1.3572642 -0.09607223 -1.249216010
#> [48,]  0.3873464  0.41138065 -0.001234302
#> [49,]  2.9278013  2.67791507  3.629988512
#> [50,] -2.4898856 -2.53479214 -3.853167277
#> 
#> $meta
#>              [,1]        [,2]       [,3]
#>  [1,] -0.59563267 -0.90002433 -0.8795599
#>  [2,]  2.67095768  2.47810147  2.5808935
#>  [3,]  3.45840247  3.30020002  3.3891269
#>  [4,] -1.46188457 -1.61687371 -1.8540286
#>  [5,]  1.63993174  1.62517210  1.4469087
#>  [6,] -5.46260433 -5.55647351 -6.0376931
#>  [7,]  2.55669035  2.49244482  2.4744017
#>  [8,]  3.95001353  3.95357827  3.9602483
#>  [9,]  2.67230496  2.60445934  2.5473784
#> [10,] -0.06333041 -0.24056875 -0.3260788
#> [11,] -3.08902202 -3.15794007 -3.4647823
#> [12,]  1.14770503  0.68919861  0.8998934
#> [13,]  1.01616716  0.90735085  0.7856638
#> [14,] -4.57317204 -4.32126018 -5.0121129
#> [15,]  1.66901306  1.65576325  1.5602187
#> [16,]  1.30059629  0.97347727  1.1517702
#> [17,]  2.54722966  2.37283946  2.4500803
#> [18,] -0.83871748 -1.15700321 -1.2025308
#> [19,] -3.35980259 -3.47252937 -3.7627773
#> [20,]  1.14619325  1.05179463  0.9069345
#> [21,]  1.10144529  1.11690231  0.9085406
#> [22,]  1.01964166  0.62895804  0.7479229
#> [23,] -2.39597908 -2.66035613 -2.7942077
#> [24,] -3.27011549 -3.34669753 -3.6958912
#> [25,]  1.86402073  1.54195928  1.6734630
#> [26,] -5.07189459 -4.90372974 -5.5570170
#> [27,] -3.04298327 -3.13330855 -3.4259584
#> [28,] -2.29854182 -2.26246106 -2.6930823
#> [29,]  0.06371202 -0.04398019 -0.1212793
#> [30,] -0.50607066 -0.50300677 -0.8264913
#> [31,] -1.10152893 -1.31322625 -1.3820411
#> [32,] -2.94579505 -2.81601879 -3.3576427
#> [33,]  0.60928985  0.33633488  0.4268274
#> [34,]  3.60951584  3.21244112  3.5572606
#> [35,] -0.62221676 -0.80139080 -0.8062647
#> [36,]  1.15893179  1.05945946  1.0229827
#> [37,]  1.86880093  1.62188578  1.7719707
#> [38,] -0.56005416 -0.56280918 -0.8874085
#> [39,] -0.57012886 -0.82870684 -0.8213634
#> [40,]  2.69250620  2.75566901  2.5358990
#> [41,]  0.65637153  0.19064789  0.4769644
#> [42,] -2.70523844 -2.61931747 -3.0558713
#> [43,]  2.13395943  1.81842962  1.9202530
#> [44,]  1.08933714  1.32199345  0.8918323
#> [45,]  0.87872139  0.60266852  0.6715000
#> [46,]  6.58840317  6.27388617  6.6620334
#> [47,] -0.03083230  0.02072140 -0.1677526
#> [48,]  0.36265531  0.28827976  0.1348278
#> [49,]  3.62457049  3.32176166  3.5519022
#> [50,] -3.30474852 -3.23086153 -3.7021257
#>