Make Predictions

Predicts outcome from features with stacked model.

# S3 method for joinet
predict(object, newx, type = "response", ...)

Arguments

object

joinet object

newx

covariates: numeric matrix with \(n\) rows (samples) and \(p\) columns (variables)

type

character "link" or "response"

...

further arguments (not applicable)

Value

This function returns predictions from base and meta learners. The slots base and meta each contain a matrix with \(n\) rows (samples) and \(q\) columns (variables).

Examples

# \dontshow{
if(!grepl('SunOS',Sys.info()['sysname'])){
n <- 50; p <- 100; q <- 3
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
Y <- replicate(n=q,expr=rnorm(n=n,mean=rowSums(X[,1:5])))
Y[,1] <- 1*(Y[,1]>median(Y[,1]))
object <- joinet(Y=Y,X=X,family=c("binomial","gaussian","gaussian"))
predict(object,newx=X)}# }
#> $base
#>            [,1]       [,2]        [,3]
#>  [1,] 0.6407102  3.1680783  2.85029995
#>  [2,] 0.4529690 -2.8823638 -3.03523395
#>  [3,] 0.8623641  1.7979907  2.23813064
#>  [4,] 0.7201645  1.1777996  1.44722947
#>  [5,] 0.5584302  0.3902722 -0.42092527
#>  [6,] 0.7773234  3.5366003  4.01188203
#>  [7,] 0.3169053 -2.8196259 -3.60681808
#>  [8,] 0.2078633 -0.8938234 -1.31034283
#>  [9,] 0.1963576 -1.6802347 -0.78831030
#> [10,] 0.7106076 -0.3952090 -1.52924369
#> [11,] 0.6059456 -1.4198203 -0.10011598
#> [12,] 0.2136019 -0.3695911 -0.69855770
#> [13,] 0.1373861 -2.4383394 -2.79752459
#> [14,] 0.6913715  0.3438205  0.79712312
#> [15,] 0.8481990 -0.1190671  0.12091095
#> [16,] 0.2343026 -0.5352938 -1.04936790
#> [17,] 0.3032167 -0.8660165 -1.11884826
#> [18,] 0.5394759 -0.5687872 -1.02673968
#> [19,] 0.5188516 -0.6643618  0.24627885
#> [20,] 0.4995880 -0.2132547 -0.08779512
#> [21,] 0.7455507  1.0971569  2.37937515
#> [22,] 0.3281272 -0.9157235 -1.41535599
#> [23,] 0.8016874  1.4557081  0.87495776
#> [24,] 0.7170884  0.4343228  1.06522705
#> [25,] 0.8291823  1.8817787  2.13704911
#> [26,] 0.7825912  2.3381396  2.49105383
#> [27,] 0.8565318  1.5849111  1.55928598
#> [28,] 0.3871362 -0.7318956  0.04224873
#> [29,] 0.5616469  1.1960311  1.37996092
#> [30,] 0.2790156 -2.3278303 -1.48748611
#> [31,] 0.3403796 -0.6176256 -0.89330793
#> [32,] 0.2473678 -4.5789527 -4.68424487
#> [33,] 0.4854308  0.6377028 -1.12350999
#> [34,] 0.2080084 -0.9734144 -1.29259414
#> [35,] 0.1571594 -2.5762407 -3.73615456
#> [36,] 0.2867933 -1.6045284 -2.33635451
#> [37,] 0.4242444  0.1033442  0.55647081
#> [38,] 0.1372867  0.1821131 -1.53689954
#> [39,] 0.6185643  1.2869929  1.77441048
#> [40,] 0.5823634 -0.3139989  1.33845258
#> [41,] 0.7426366  2.4221123  1.13533109
#> [42,] 0.4515816 -2.5700485 -2.39170486
#> [43,] 0.1611646  0.8925041  0.35456691
#> [44,] 0.3367430 -0.7537727 -1.74324659
#> [45,] 0.8149929  2.9670664  2.43856760
#> [46,] 0.2911118 -3.2092829 -3.08080353
#> [47,] 0.5497539 -0.3453375 -1.54784731
#> [48,] 0.4079904 -0.6888655 -0.43015083
#> [49,] 0.7081383 -0.2198393  1.16212714
#> [50,] 0.7240975  1.0255780  2.88215216
#> 
#> $meta
#>               [,1]        [,2]        [,3]
#>  [1,] 0.9924925674  2.97175982  3.78554678
#>  [2,] 0.0183184016 -3.01278430 -3.54852212
#>  [3,] 0.9888690749  2.08533949  3.16253079
#>  [4,] 0.9426171300  1.34985770  1.95842693
#>  [5,] 0.6168360252 -0.14601924  0.07132432
#>  [6,] 0.9983799979  3.85869919  4.98649391
#>  [7,] 0.0085384084 -3.36225309 -4.10604141
#>  [8,] 0.0875359043 -1.18215257 -1.84889064
#>  [9,] 0.0633722924 -1.12278009 -1.99745285
#> [10,] 0.4360124626 -1.14688242 -0.77607603
#> [11,] 0.3621362126 -0.58230768 -0.66390028
#> [12,] 0.1764696463 -0.59738081 -1.16396534
#> [13,] 0.0077905722 -2.69993428 -3.83768752
#> [14,] 0.8385482803  0.62978716  1.03309518
#> [15,] 0.8362071116  0.02496328  0.73661418
#> [16,] 0.1439135748 -0.88463183 -1.42345203
#> [17,] 0.1428744888 -1.04767427 -1.50534101
#> [18,] 0.3331448950 -0.88184634 -0.87097793
#> [19,] 0.4839434098 -0.08778472 -0.17458824
#> [20,] 0.5045575528 -0.14436437 -0.15969750
#> [21,] 0.9679530977  1.92749715  2.54140429
#> [22,] 0.1284176630 -1.25826869 -1.66797403
#> [23,] 0.9542224700  1.07661209  1.94901941
#> [24,] 0.8788008968  0.83644085  1.30025508
#> [25,] 0.9863580031  2.04944018  3.04123926
#> [26,] 0.9900708732  2.44233781  3.39086576
#> [27,] 0.9795956548  1.56781572  2.60488039
#> [28,] 0.3365371601 -0.24457320 -0.56014799
#> [29,] 0.9002545110  1.31259484  1.63766081
#> [30,] 0.0377060764 -1.80835808 -2.60009204
#> [31,] 0.2102500351 -0.81244934 -1.15648363
#> [32,] 0.0008682693 -4.69002550 -5.89852792
#> [33,] 0.5002263355 -0.51489536 -0.34459581
#> [34,] 0.0836275346 -1.19896695 -1.88217777
#> [35,] 0.0045281099 -3.35966432 -4.42777844
#> [36,] 0.0405354735 -2.10282944 -2.70195644
#> [37,] 0.5996068391  0.38755181  0.28850482
#> [38,] 0.1129155963 -0.94614890 -1.59335613
#> [39,] 0.9371790366  1.60160194  2.02945095
#> [40,] 0.7462824189  0.74753299  0.80128401
#> [41,] 0.9750146849  1.59034186  2.51085785
#> [42,] 0.0337338460 -2.48287375 -2.97940787
#> [43,] 0.4471412220  0.53740711  0.04816391
#> [44,] 0.1238058246 -1.41385934 -1.76314118
#> [45,] 0.9946302554  2.63229653  3.79470333
#> [46,] 0.0078711937 -3.15891713 -4.05088832
#> [47,] 0.3093158562 -1.14121395 -1.04964581
#> [48,] 0.2972608331 -0.53654337 -0.79093172
#> [49,] 0.8174403490  0.66638205  0.97534030
#> [50,] 0.9725366061  2.22905562  2.76533833
#> 
if (FALSE) {
n <- 50; p <- 100; q <- 3
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
Y <- replicate(n=q,expr=rnorm(n=n,mean=rowSums(X[,1:5])))
Y[,1] <- 1*(Y[,1]>median(Y[,1]))
object <- joinet(Y=Y,X=X,family=c("binomial","gaussian","gaussian"))
predict(object,newx=X)}