Make Predictions

Predicts outcome from features with stacked model.

# S3 method for joinet
predict(object, newx, type = "response", ...)

Arguments

object
joinet object

newx
covariates: numeric matrix with \(n\) rows (samples) and \(p\) columns (variables)

type
character "link" or "response"

...
further arguments (not applicable)

Value

This function returns predictions from base and meta learners. The slots base and meta each contain a matrix with \(n\) rows (samples) and \(q\) columns (variables).

Examples

n <- 50; p <- 100; q <- 3
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
Y <- replicate(n=q,expr=rnorm(n=n,mean=rowSums(X[,1:5])))
Y[,1] <- 1*(Y[,1]>median(Y[,1]))
object <- joinet(Y=Y,X=X,family=c("binomial","gaussian","gaussian"))
predict(object,newx=X)

#> $base
#>             [,1]         [,2]        [,3]
#>  [1,] 0.41837245 -1.928585385 -2.49734037
#>  [2,] 0.73430703 -1.232615541 -1.92258279
#>  [3,] 0.89830295  2.189899197  1.90525393
#>  [4,] 0.89750302  2.620239322  3.16623238
#>  [5,] 0.75114829  0.007031193 -0.17521804
#>  [6,] 0.14470451 -2.282808008 -2.29598057
#>  [7,] 0.09766868 -1.173475809 -1.01900217
#>  [8,] 0.90945310  3.743238898  3.54799601
#>  [9,] 0.80093334  0.020539090  0.66028304
#> [10,] 0.41399970 -2.478399066 -2.17809084
#> [11,] 0.87769011  3.710699246  3.43305524
#> [12,] 0.10649510 -0.594219557 -0.26688072
#> [13,] 0.07607328 -4.684823994 -4.25280223
#> [14,] 0.20653264  0.507040407  0.46010982
#> [15,] 0.22855559 -0.707175402 -0.50492476
#> [16,] 0.77487110 -1.108208462 -0.97986741
#> [17,] 0.22673908 -1.104045322 -0.53058056
#> [18,] 0.88702215 -0.008662017  0.05825129
#> [19,] 0.44743009  1.637425052  0.93757078
#> [20,] 0.39800750 -2.875874417 -3.01614676
#> [21,] 0.16726717 -0.589831578 -0.90077905
#> [22,] 0.64835839 -0.667092621 -0.74610665
#> [23,] 0.75523423 -0.829522996  0.17771580
#> [24,] 0.11655731 -0.337081023  0.12605514
#> [25,] 0.08593514 -3.783144741 -3.78084340
#> [26,] 0.60801411  1.738083231  2.02830016
#> [27,] 0.31502035 -1.667382334 -2.04750083
#> [28,] 0.91002966  2.406116563  1.79844125
#> [29,] 0.08608203 -3.935505412 -3.55307761
#> [30,] 0.16252154 -3.458884133 -4.05879847
#> [31,] 0.78128831  1.328059672  0.88936337
#> [32,] 0.83275586  0.090645605  0.79816444
#> [33,] 0.57741370  0.248453011  0.72755333
#> [34,] 0.30918554 -0.226150688 -0.24999979
#> [35,] 0.18514098 -2.839059766 -3.48961345
#> [36,] 0.77872138  1.585662558  1.89663933
#> [37,] 0.31999646 -1.471891490 -2.30443277
#> [38,] 0.23445394 -1.237596437 -2.02415725
#> [39,] 0.93005087  2.285273037  2.71846176
#> [40,] 0.73774300  0.642560288  1.00153549
#> [41,] 0.17295702 -1.548693979 -0.99259618
#> [42,] 0.75888127  0.944554378  0.64050854
#> [43,] 0.60958826 -0.254105578 -0.54161539
#> [44,] 0.38532271 -1.029882728 -2.25850892
#> [45,] 0.62974890 -0.089590037 -0.18848433
#> [46,] 0.84637050  1.692734994  1.24297687
#> [47,] 0.40207150 -0.545823224 -1.07908874
#> [48,] 0.25251723 -2.482347877 -2.62104545
#> [49,] 0.77840583  0.140319746 -0.48219207
#> [50,] 0.32655598 -1.650189753 -0.59297402
#> 
#> $meta
#>             [,1]       [,2]        [,3]
#>  [1,] 0.10220137 -2.8366810 -2.58549073
#>  [2,] 0.22806673 -1.8261107 -1.76560859
#>  [3,] 0.95484141  2.8130533  2.62981542
#>  [4,] 0.98730517  4.1550932  3.46554651
#>  [5,] 0.64845882  0.1406465 -0.04254334
#>  [6,] 0.08701080 -3.0352129 -2.82347419
#>  [7,] 0.24075871 -1.6987418 -1.38562725
#>  [8,] 0.99171832  4.7229217  4.59039026
#>  [9,] 0.82540857  1.0679076  0.27264687
#> [10,] 0.13617019 -2.5849863 -2.95244727
#> [11,] 0.98981676  4.5178102  4.52008410
#> [12,] 0.41469313 -0.8318596 -0.60385413
#> [13,] 0.01027685 -5.5214546 -5.64340291
#> [14,] 0.64930175  0.2459392  0.62723142
#> [15,] 0.41386111 -0.8637077 -0.78947614
#> [16,] 0.45380923 -0.7919286 -1.31410010
#> [17,] 0.40677141 -0.9432038 -1.14733110
#> [18,] 0.75190956  0.6145579  0.02843515
#> [19,] 0.80419429  1.1601342  1.79330367
#> [20,] 0.06109382 -3.5090532 -3.60577442
#> [21,] 0.29707148 -1.3498797 -0.83013621
#> [22,] 0.47255039 -0.6514714 -0.84183516
#> [23,] 0.72776624  0.4000683 -0.64909323
#> [24,] 0.52225525 -0.3720559 -0.23537421
#> [25,] 0.01722845 -4.8908178 -4.68018275
#> [26,] 0.93788800  2.4494650  2.27769155
#> [27,] 0.13861339 -2.4543910 -2.19278248
#> [28,] 0.95152342  2.7650921  2.78092452
#> [29,] 0.02171825 -4.6769505 -4.73129956
#> [30,] 0.01570247 -4.9537324 -4.49632044
#> [31,] 0.85305109  1.4396438  1.50410997
#> [32,] 0.85232458  1.2707274  0.38427369
#> [33,] 0.79166183  0.8978043  0.49722869
#> [34,] 0.50644688 -0.4392424 -0.27447864
#> [35,] 0.02911065 -4.2534806 -3.74533419
#> [36,] 0.94251258  2.4986830  2.09603282
#> [37,] 0.11040807 -2.6864056 -2.11437644
#> [38,] 0.12885189 -2.4765109 -1.80685937
#> [39,] 0.98204701  3.7581482  3.00880873
#> [40,] 0.85949146  1.4074652  0.94281190
#> [41,] 0.27978357 -1.5562501 -1.70556618
#> [42,] 0.81267168  1.1047951  1.07696011
#> [43,] 0.51433013 -0.4308933 -0.40489984
#> [44,] 0.12318913 -2.5118215 -1.70952119
#> [45,] 0.60957965 -0.0276023 -0.13221472
#> [46,] 0.90379982  1.9552732  1.95275535
#> [47,] 0.32662397 -1.2267246 -0.85622205
#> [48,] 0.07567690 -3.2214496 -3.11672913
#> [49,] 0.58299820 -0.1185430 -0.03693255
#> [50,] 0.42367190 -0.9519434 -1.64962005
#>