Make Predictions

Predicts outcome from features with stacked model.

# S3 method for joinet
predict(object, newx, type = "response", ...)

Arguments

object
joinet object

newx
covariates: numeric matrix with \(n\) rows (samples) and \(p\) columns (variables)

type
character "link" or "response"

...
further arguments (not applicable)

Value

This function returns predictions from base and meta learners. The slots base and meta each contain a matrix with \(n\) rows (samples) and \(q\) columns (variables).

Examples

n <- 50; p <- 100; q <- 3
X <- matrix(rnorm(n*p),nrow=n,ncol=p)
Y <- replicate(n=q,expr=rnorm(n=n,mean=rowSums(X[,1:5])))
Y[,1] <- 1*(Y[,1]>median(Y[,1]))
object <- joinet(Y=Y,X=X,family=c("binomial","gaussian","gaussian"))
predict(object,newx=X)

#> $base
#>             [,1]         [,2]        [,3]
#>  [1,] 0.41837245 -1.928585385 -2.49734037
#>  [2,] 0.73430703 -1.232615541 -1.92258279
#>  [3,] 0.89830295  2.189899197  1.90525393
#>  [4,] 0.89750302  2.620239322  3.16623238
#>  [5,] 0.75114829  0.007031193 -0.17521804
#>  [6,] 0.14470451 -2.282808008 -2.29598057
#>  [7,] 0.09766868 -1.173475809 -1.01900217
#>  [8,] 0.90945310  3.743238898  3.54799601
#>  [9,] 0.80093334  0.020539090  0.66028304
#> [10,] 0.41399970 -2.478399066 -2.17809084
#> [11,] 0.87769011  3.710699246  3.43305524
#> [12,] 0.10649510 -0.594219557 -0.26688072
#> [13,] 0.07607328 -4.684823994 -4.25280223
#> [14,] 0.20653264  0.507040407  0.46010982
#> [15,] 0.22855559 -0.707175402 -0.50492476
#> [16,] 0.77487110 -1.108208462 -0.97986741
#> [17,] 0.22673908 -1.104045322 -0.53058056
#> [18,] 0.88702215 -0.008662017  0.05825129
#> [19,] 0.44743009  1.637425052  0.93757078
#> [20,] 0.39800750 -2.875874417 -3.01614676
#> [21,] 0.16726717 -0.589831578 -0.90077905
#> [22,] 0.64835839 -0.667092621 -0.74610665
#> [23,] 0.75523423 -0.829522996  0.17771580
#> [24,] 0.11655731 -0.337081023  0.12605514
#> [25,] 0.08593514 -3.783144741 -3.78084340
#> [26,] 0.60801411  1.738083231  2.02830016
#> [27,] 0.31502035 -1.667382334 -2.04750083
#> [28,] 0.91002966  2.406116563  1.79844125
#> [29,] 0.08608203 -3.935505412 -3.55307761
#> [30,] 0.16252154 -3.458884133 -4.05879847
#> [31,] 0.78128831  1.328059672  0.88936337
#> [32,] 0.83275586  0.090645605  0.79816444
#> [33,] 0.57741370  0.248453011  0.72755333
#> [34,] 0.30918554 -0.226150688 -0.24999979
#> [35,] 0.18514098 -2.839059766 -3.48961345
#> [36,] 0.77872138  1.585662558  1.89663933
#> [37,] 0.31999646 -1.471891490 -2.30443277
#> [38,] 0.23445394 -1.237596437 -2.02415725
#> [39,] 0.93005087  2.285273037  2.71846176
#> [40,] 0.73774300  0.642560288  1.00153549
#> [41,] 0.17295702 -1.548693979 -0.99259618
#> [42,] 0.75888127  0.944554378  0.64050854
#> [43,] 0.60958826 -0.254105578 -0.54161539
#> [44,] 0.38532271 -1.029882728 -2.25850892
#> [45,] 0.62974890 -0.089590037 -0.18848433
#> [46,] 0.84637050  1.692734994  1.24297687
#> [47,] 0.40207150 -0.545823224 -1.07908874
#> [48,] 0.25251723 -2.482347877 -2.62104545
#> [49,] 0.77840583  0.140319746 -0.48219207
#> [50,] 0.32655598 -1.650189753 -0.59297402
#> 
#> $meta
#>              [,1]        [,2]        [,3]
#>  [1,] 0.120026857 -2.63552370 -2.58637745
#>  [2,] 0.311517089 -1.57742790 -1.80530334
#>  [3,] 0.970038575  2.83530787  2.50770389
#>  [4,] 0.988790465  3.94203727  3.41749124
#>  [5,] 0.702397389  0.20191699 -0.07301232
#>  [6,] 0.069874337 -3.01383050 -2.75113316
#>  [7,] 0.178293600 -1.79055373 -1.33607264
#>  [8,] 0.994417547  4.65204823  4.42063500
#>  [9,] 0.825454272  0.92712611  0.33068070
#> [10,] 0.122295626 -2.58218395 -2.83894610
#> [11,] 0.992972765  4.45732386  4.34260769
#> [12,] 0.315048704 -0.99259865 -0.55695917
#> [13,] 0.006376865 -5.52584767 -5.43421552
#> [14,] 0.604946394  0.15758601  0.59296195
#> [15,] 0.356113893 -0.95306415 -0.75195663
#> [16,] 0.496057561 -0.75220207 -1.26978360
#> [17,] 0.320794509 -1.10958000 -1.05497247
#> [18,] 0.804808814  0.64766735  0.02552338
#> [19,] 0.834600977  1.22772641  1.64662367
#> [20,] 0.062637571 -3.37591257 -3.52530885
#> [21,] 0.271715919 -1.32688363 -0.85251362
#> [22,] 0.507637975 -0.60258988 -0.83624877
#> [23,] 0.694579538  0.19562106 -0.51989858
#> [24,] 0.406609311 -0.57645076 -0.18326513
#> [25,] 0.012589218 -4.82289140 -4.55079851
#> [26,] 0.935270035  2.28086412  2.23449684
#> [27,] 0.143615352 -2.33101536 -2.18288391
#> [28,] 0.971697981  2.86526752  2.61585773
#> [29,] 0.013944811 -4.69966856 -4.55513448
#> [30,] 0.015768781 -4.71928228 -4.44410746
#> [31,] 0.891631228  1.50966480  1.39712353
#> [32,] 0.853434630  1.11696980  0.44705866
#> [33,] 0.772218415  0.74754148  0.52945800
#> [34,] 0.480168883 -0.47818857 -0.27919160
#> [35,] 0.030260125 -4.02999089 -3.72147791
#> [36,] 0.946608961  2.36236767  2.06071075
#> [37,] 0.131953701 -2.45489591 -2.16067825
#> [38,] 0.142539577 -2.28403264 -1.85682897
#> [39,] 0.985751175  3.60577123  2.96090103
#> [40,] 0.863879903  1.29591149  0.94749069
#> [41,] 0.204237405 -1.70922989 -1.59867884
#> [42,] 0.852342486  1.15387209  0.99870566
#> [43,] 0.559959816 -0.35427152 -0.43712903
#> [44,] 0.170830538 -2.18882713 -1.81502079
#> [45,] 0.639682562 -0.00439300 -0.15009278
#> [46,] 0.934408843  2.02660474  1.83111336
#> [47,] 0.358655524 -1.10381588 -0.90429501
#> [48,] 0.070055480 -3.13250260 -3.05061558
#> [49,] 0.681260321  0.05548031 -0.11998977
#> [50,] 0.317122867 -1.19819266 -1.48472862
#>